pivoting to web app, not mobile app
there's no need for this to be a mobile app other than the fact
that most apps that do this sort of thing are mobile apps. but the
problem that's being solved doesn't require it to be a mobile app.
this can be done via a mobile-friendly web app. and since i plan
very little frills in terms of design, a web app should be perfectly
sufficient